sources: Reorganize from one long file into several shorter files
I found it was starting to get harder to find your way around and ignore
adjacent things that weren't relevant.

No functional code changes; only added lines which aren't identical
copies of removed lines are for require() path adjustments.
