--- src/sources.js
@@ -1,835 +1,0 @@
-/* eslint-disable no-use-before-define */
-const AWS = require("aws-sdk");
-const zlib = require("zlib");
-const yamlFront = require("yaml-front-matter");
-const {fetch} = require("./fetch");
-const queryString = require("query-string");
-const {NotFound} = require('http-errors');
-const {NoResourcePathError} = require("./exceptions");
-const utils = require("./utils");
-
-const S3 = new AWS.S3();
-
-/* These Source, Dataset, and Narrative classes contain information to map an
- * array of dataset/narrative path parts onto a URL.  Source selection and
- * dataset path aliasing (/flu → /flu/seasonal/h3n2/ha/3y) is handled in
- * utils/prefix.parsePrefix().
- *
- * The class definitions would be a bit shorter/prettier if we were using Babel
- * to allow class properties on Node.
- */
-
-class Source {
-  static get _name() {
-    throw new Error("_name() must be implemented by subclasses");
-  }
-  get name() {
-    return this.constructor._name;
-  }
-  async baseUrl() {
-    throw new Error("async baseUrl() must be implemented by subclasses");
-  }
-  async urlFor(path, method = 'GET') { // eslint-disable-line no-unused-vars
-    const url = new URL(path, await this.baseUrl());
-    return url.toString();
-  }
-  static isGroup() { /* is the source a "nextstrain group"? */
-    return false;
-  }
-  dataset(pathParts) {
-    return new Dataset(this, pathParts);
-  }
-  narrative(pathParts) {
-    return new Narrative(this, pathParts);
-  }
-
-  // eslint-disable-next-line no-unused-vars
-  secondTreeOptions(path) {
-    return [];
-  }
-
-  availableDatasets() {
-    return [];
-  }
-  availableNarratives() {
-    return [];
-  }
-
-  /* Static access control for this entire source, regardless of any
-   * instance-specific parameters.
-   */
-  static visibleToUser(user) { // eslint-disable-line no-unused-vars
-    return true;
-  }
-
-  /* Instance-specific access control delegates to the static method by
-   * default.
-   */
-  visibleToUser(user) {
-    return this.constructor.visibleToUser(user);
-  }
-
-  async getInfo() {
-    throw new Error("getInfo() must be implemented by subclasses");
-  }
-}
-
-class Resource {
-  constructor(source, pathParts) {
-    this.source = source;
-    this.pathParts = pathParts;
-
-    // Require baseParts, otherwise we have no actual dataset/narrative path.
-    // This inspects baseParts because some of the pathParts (above) may not
-    // apply, which each Dataset/Narrative subclass determines for itself.
-    if (!this.baseParts.length) {
-      throw new NoResourcePathError();
-    }
-  }
-  get baseParts() {
-    return this.pathParts.slice();
-  }
-  get baseName() {
-    return this.baseParts.join("_");
-  }
-  async exists() {
-    throw new Error("exists() must be implemented by Resource subclasses");
-  }
-  subresource(type) { // eslint-disable-line no-unused-vars
-    throw new Error("subresource() must be implemented by Resource subclasses");
-  }
-}
-
-class Subresource {
-  constructor(resource, type) {
-    if (this.constructor === Subresource) {
-      throw new Error("Subresource interface class must be subclassed");
-    }
-    if (!(resource instanceof Resource)) {
-      throw new Error(`invalid Subresource parent resource type: ${resource.constructor}`);
-    }
-    if (!this.constructor.validTypes.includes(type)) {
-      throw new Error(`invalid Subresource type: ${type}`);
-    }
-    this.resource = resource;
-    this.type = type;
-  }
-  static get validTypes() {
-    throw new Error("validTypes() must be implemented by Subresource subclasses");
-  }
-  async url(method = 'GET') {
-    return await this.resource.source.urlFor(this.baseName, method);
-  }
-  get baseName() {
-    throw new Error("baseName() must be implemented by Subresource subclasses");
-  }
-}
-
-class DatasetSubresource extends Subresource {
-  static validTypes = ["main", "root-sequence", "tip-frequencies", "meta", "tree"];
-
-  get baseName() {
-    return this.type === "main"
-      ? `${this.resource.baseName}.json`
-      : `${this.resource.baseName}_${this.type}.json`;
-  }
-}
-
-class NarrativeSubresource extends Subresource {
-  static validTypes = ["md"];
-
-  get baseName() {
-    return `${this.resource.baseName}.md`;
-  }
-}
-
-class Dataset extends Resource {
-  async exists() {
-    const method = "HEAD";
-    const _exists = async (type) =>
-      (await fetch(await this.subresource(type).url(method), {method, cache: "no-store"})).status === 200;
-
-    const all = async (...promises) =>
-      (await Promise.all(promises)).every(x => x);
-
-    return (await _exists("main"))
-        || (await all(_exists("meta"), _exists("tree")))
-        || false;
-  }
-
-  /**
-   * Resolve this Dataset to its full canonical path, if this one is partially
-   * specified and defaults exist (i.e. if this one is an alias).
-   *
-   * For example, in our core source, /flu/seasonal/h3n2 is an alias for
-   * /flu/seasonal/h3n2/ha/2y.
-   *
-   * Returns this Dataset itself if it's already the canonical one or no
-   * aliases exist.  Thus, you can compare `dataset === dataset.resolve()` to
-   * see if `dataset` is an alias.
-   *
-   * @returns {Dataset}
-   */
-  resolve() {
-    /* XXX TODO: Reimplement this in terms of methods on the source, not by
-     * breaking encapsulation by using a process-wide global.
-     *   -trs, 26 Oct 2021 (based on a similar comment 5 Sept 2019)
-     */
-    const sourceName = this.source.name;
-    const prefixParts = this.pathParts;
-
-    if (!global.availableDatasets[sourceName]) {
-      utils.verbose("Can't compare against available datasets as there are none!");
-      return this;
-    }
-
-    const doesPathExist = (pathToCheck) =>
-      global.availableDatasets[sourceName]
-        .includes(pathToCheck);
-
-    const prefix = prefixParts.join("/");
-
-    if (doesPathExist(prefix)) {
-      return this;
-    }
-
-    /* if we are here, then the path doesn't match any available datasets exactly */
-    const nextDefaultPart = global.availableDatasets.defaults[sourceName][prefix];
-
-    if (nextDefaultPart) {
-      const dataset = new this.constructor(this.source, [...prefixParts, nextDefaultPart]);
-      return dataset.resolve();
-    }
-
-    return this;
-  }
-
-  get isRequestValidWithoutDataset() {
-    return false;
-  }
-
-  subresource(type) {
-    return new DatasetSubresource(this, type);
-  }
-}
-
-class Narrative extends Resource {
-  async exists() {
-    const method = "HEAD";
-    const _exists = async () =>
-      (await fetch(await this.subresource("md").url(method), {method, cache: "no-store"})).status === 200;
-
-    return (await _exists()) || false;
-  }
-
-  subresource(type) {
-    return new NarrativeSubresource(this, type);
-  }
-}
-
-class CoreSource extends Source {
-  static get _name() { return "core"; }
-  async baseUrl() { return "http://data.nextstrain.org/"; }
-  get repo() { return "nextstrain/narratives"; }
-  get branch() { return "master"; }
-
-  async urlFor(path, method = 'GET') { // eslint-disable-line no-unused-vars
-    const baseUrl = path.endsWith(".md")
-      ? `https://raw.githubusercontent.com/${this.repo}/${await this.branch}/`
-      : await this.baseUrl();
-
-    const url = new URL(path, baseUrl);
-    return url.toString();
-  }
-
-  // The computation of these globals should move here.
-  secondTreeOptions(path) {
-    return (global.availableDatasets.secondTreeOptions[this.name] || {})[path] || [];
-  }
-
-  availableDatasets() {
-    return global.availableDatasets[this.name] || [];
-  }
-
-  async availableNarratives() {
-    const qs = queryString.stringify({ref: this.branch});
-    const response = await fetch(`https://api.github.com/repos/${this.repo}/contents?${qs}`);
-
-    if (response.status === 404) throw new NotFound();
-    else if (response.status !== 200 && response.status !== 304) {
-      utils.warn(`Error fetching available narratives from GitHub for source ${this.name}`, await utils.responseDetails(response));
-      return [];
-    }
-
-    const files = await response.json();
-    return files
-      .filter((file) => file.type === "file")
-      .filter((file) => file.name !== "README.md")
-      .filter((file) => file.name.endsWith(".md"))
-      .map((file) => file.name
-        .replace(/[.]md$/, "")
-        .split("_")
-        .join("/"));
-  }
-
-  async getInfo() {
-    return {
-      title: `Nextstrain ${this.name} datasets & narratives`,
-      showDatasets: true,
-      showNarratives: true,
-    };
-  }
-}
-
-class CoreStagingSource extends CoreSource {
-  static get _name() { return "staging"; }
-  async baseUrl() { return "http://staging.nextstrain.org/"; }
-  get repo() { return "nextstrain/narratives"; }
-  get branch() { return "staging"; }
-}
-
-class CommunitySource extends Source {
-  constructor(owner, repoName) {
-    super();
-
-    // The GitHub owner and repo names are required.
-    if (!owner) throw new Error(`Cannot construct a ${this.constructor.name} without an owner`);
-    if (!repoName) throw new Error(`Cannot construct a ${this.constructor.name} without a repoName`);
-
-    this.owner = owner;
-    [this.repoName, this.branch] = repoName.split(/@/, 2);
-    this.branchExplicitlyDefined = !!this.branch;
-
-    if (!this.repoName) throw new Error(`Cannot construct a ${this.constructor.name} without a repoName after splitting on /@/`);
-
-    this.defaultBranch = fetch(`https://api.github.com/repos/${this.owner}/${this.repoName}`)
-      .then((res) => res.json())
-      .then((data) => data.default_branch)
-      .catch(() => {
-        console.log(`Error interpreting the default branch of ${this.constructor.name} for ${this.owner}/${this.repoName}`);
-        return "master";
-      });
-    if (!this.branch) {
-      this.branch = this.defaultBranch;
-    }
-  }
-
-  static get _name() { return "community"; }
-  get repo() { return `${this.owner}/${this.repoName}`; }
-  async baseUrl() {
-    return `https://github.com/${this.repo}/raw/${await this.branch}/`;
-  }
-
-  async repoNameWithBranch() {
-    const branch = await this.branch;
-    const defaultBranch = await this.defaultBranch;
-    if (branch === defaultBranch && !this.branchExplicitlyDefined) {
-      return this.repoName;
-    }
-    return `${this.repoName}@${branch}`;
-  }
-
-  dataset(pathParts) {
-    return new CommunityDataset(this, pathParts);
-  }
-  narrative(pathParts) {
-    return new CommunityNarrative(this, pathParts);
-  }
-
-  async availableDatasets() {
-    const qs = queryString.stringify({ref: await this.branch});
-    const response = await fetch(`https://api.github.com/repos/${this.repo}/contents/auspice?${qs}`);
-
-    if (response.status === 404) throw new NotFound();
-    else if (response.status !== 200 && response.status !== 304) {
-      utils.warn(`Error fetching available datasets from GitHub for source ${this.name}`, await utils.responseDetails(response));
-      return [];
-    }
-
-    const filenames = (await response.json())
-      .filter((file) => file.type === "file")
-      // remove anything which doesn't start with the repo name, which is required of community datasets
-      .filter((file) => file.name.startsWith(this.repoName))
-      .map((file) => file.name);
-    const pathnames = utils.getDatasetsFromListOfFilenames(filenames)
-      // strip out the repo name from the start of the pathnames
-      // as CommunityDataset().baseParts will add this in
-      .map((pathname) => pathname.replace(`${this.repoName}/`, ""));
-    return pathnames;
-  }
-
-  async availableNarratives() {
-    const qs = queryString.stringify({ref: await this.branch});
-    const response = await fetch(`https://api.github.com/repos/${this.repo}/contents/narratives?${qs}`);
-
-    if (response.status !== 200 && response.status !== 304) {
-      if (response.status !== 404) {
-        // not found doesn't warrant an error print, it means there are no narratives for this repo
-        utils.warn(`Error fetching available narratives from GitHub for source ${this.name}`, await utils.responseDetails(response));
-      }
-      return [];
-    }
-
-    const files = await response.json();
-    return files
-      .filter((file) => file.type === "file")
-      .filter((file) => file.name !== "README.md")
-      .filter((file) => file.name.endsWith(".md"))
-      .filter((file) => file.name.startsWith(this.repoName))
-      .map((file) => file.name
-        .replace(this.repoName, "")
-        .replace(/^_/, "")
-        .replace(/[.]md$/, "")
-        .split("_")
-        .join("/"));
-  }
-  async getInfo() {
-    /* could attempt to fetch a certain file from the repository if we want to implement
-    this functionality in the future */
-    const branch = await this.branch;
-    return {
-      title: `${this.owner}'s "${this.repoName}" community builds`,
-      byline: `
-        Nextstrain community builds for GitHub → ${this.owner}/${this.repoName} (${branch} branch).
-        The available datasets and narratives in this repository are listed below.
-      `,
-      website: null,
-      showDatasets: true,
-      showNarratives: true,
-      /* avatar could be fetched here & sent in base64 or similar, or a link sent. The former (or similar) has the advantage
-      of private S3 buckets working, else the client will have to make (a) an authenticated request (too much work)
-      or (b) a subsequent request to nextstrain.org/charon (why not do it at once?) */
-      avatar: `https://github.com/${this.owner}.png?size=200`
-    };
-  }
-}
-
-class CommunityDataset extends Dataset {
-  get baseParts() {
-    // We require datasets are in the auspice/ directory and include the repo
-    // name in the file basename.
-    return [`auspice/${this.source.repoName}`, ...this.pathParts];
-  }
-  get isRequestValidWithoutDataset() {
-    if (!this.pathParts.length) {
-      return true;
-    }
-    return false;
-  }
-}
-
-class CommunityNarrative extends Narrative {
-  get baseParts() {
-    // We require narratives are in the narratives/ directory and include the
-    // repo name in the file basename.
-    return [`narratives/${this.source.repoName}`, ...this.pathParts];
-  }
-}
-
-
-class UrlDefinedSource extends Source {
-  static get _name() { return "fetch"; }
-
-  constructor(authority) {
-    super();
-
-    if (!authority) throw new Error(`Cannot construct a ${this.constructor.name} without a URL authority`);
-
-    this.authority = authority;
-  }
-
-  async baseUrl() {
-    return `https://${this.authority}`;
-  }
-  dataset(pathParts) {
-    return new UrlDefinedDataset(this, pathParts);
-  }
-  narrative(pathParts) {
-    return new UrlDefinedNarrative(this, pathParts);
-  }
-
-  // available datasets & narratives are unknown when the dataset is specified by the URL
-  async availableDatasets() { return []; }
-  async availableNarratives() { return []; }
-  async getInfo() { return {}; }
-}
-
-class UrlDefinedDataset extends Dataset {
-  get baseName() {
-    return this.baseParts.join("/");
-  }
-  subresource(type) {
-    return new UrlDefinedDatasetSubresource(this, type);
-  }
-}
-
-class UrlDefinedDatasetSubresource extends DatasetSubresource {
-  get baseName() {
-    const type = this.type;
-    const baseName = this.resource.baseName;
-
-    if (type === "main") {
-      return baseName;
-    }
-
-    return baseName.endsWith(".json")
-      ? `${baseName.replace(/\.json$/, '')}_${type}.json`
-      : `${baseName}_${type}`;
-  }
-}
-
-class UrlDefinedNarrative extends Narrative {
-  get baseName() {
-    return this.baseParts.join("/");
-  }
-  subresource(type) {
-    return new UrlDefinedNarrativeSubresource(this, type);
-  }
-}
-
-class UrlDefinedNarrativeSubresource extends NarrativeSubresource {
-  get baseName() {
-    return this.resource.baseName;
-  }
-}
-
-
-class S3Source extends Source {
-  get bucket() {
-    throw new Error("bucket() must be implemented by subclasses");
-  }
-  async baseUrl() {
-    return `https://${this.bucket}.s3.amazonaws.com`;
-  }
-  async _listObjects() {
-    return new Promise((resolve, reject) => {
-      let contents = [];
-      S3.listObjectsV2({Bucket: this.bucket}).eachPage((err, data, done) => {
-        if (err) {
-          utils.warn(`Could not list S3 objects for group '${this.name}'\n${err.message}`);
-          return reject(err);
-        }
-        if (data===null) { // no more data
-          return resolve(contents);
-        }
-        contents = contents.concat(data.Contents);
-        return done();
-      });
-    });
-  }
-  async availableDatasets() {
-    const objects = await this._listObjects();
-    const pathnames = utils.getDatasetsFromListOfFilenames(objects.map((object) => object.Key));
-    return pathnames;
-  }
-  async availableNarratives() {
-    // Walking logic borrowed from auspice's cli/server/getAvailable.js
-    const objects = await this._listObjects();
-    return objects
-      .map((object) => object.Key)
-      .filter((file) => file !== 'group-overview.md')
-      .filter((file) => file.endsWith(".md"))
-      .map((file) => file
-        .replace(/[.]md$/, "")
-        .split("_")
-        .join("/"));
-  }
-  async getAndDecompressObject(key) {
-    const object = await S3.getObject({ Bucket: this.bucket, Key: key}).promise();
-    if (object.ContentEncoding === 'gzip') {
-      object.Body = zlib.gunzipSync(object.Body);
-    }
-    return object.Body;
-  }
-  parseOverviewMarkdown(overviewMarkdown) {
-    const frontMatter = yamlFront.loadFront(overviewMarkdown);
-    if (!frontMatter.title) {
-      throw new Error("The overview file requires `title` in the frontmatter.");
-    }
-
-    if (frontMatter.website) {
-      if (!frontMatter.website.includes("http")) {
-        throw new Error("The website field in the overview file requires \"http\" to be present.");
-      }
-    }
-
-    if (frontMatter.showDatasets && typeof frontMatter.showDatasets !== 'boolean') {
-      throw new Error("The `showDatasets` field in the frontmatter must be a boolean.");
-    }
-
-    if (frontMatter.showNarratives && typeof frontMatter.showNarratives !== 'boolean') {
-      throw new Error("The `showNarratives` field in the frontmatter must be a boolean.");
-    }
-
-    // handle files with CRLF endings (windows)
-    const content = frontMatter.__content.replace(/\r\n/g, "\n");
-
-    return [frontMatter.title, frontMatter.byline, frontMatter.website, frontMatter.showDatasets, frontMatter.showNarratives, content];
-  }
-  /**
-   * Get information about a (particular) source.
-   * The data could be a JSON, or a markdown with YAML frontmatter. Or something else.
-   * This is very similar to our previous discussions around moving the auspice footer
-   * content to the dataset JSON. One advantage of this being outside of the auspice
-   * codebase is that we can iterate on it after pushing live to nextstrain.org
-   */
-  async getInfo() {
-    try {
-      /* attempt to fetch customisable information from S3 bucket */
-      const objects = await this._listObjects();
-      const objectKeys = objects.map((object) => object.Key);
-
-      let logoSrc;
-      if (objectKeys.includes("group-logo.png")) {
-        // Use pre-signed URL to allow client to fetch from private S3 bucket
-        logoSrc = S3.getSignedUrl('getObject', {Bucket: this.bucket, Key: "group-logo.png"});
-      }
-
-      let title = `"${this.name}" Nextstrain group`;
-      let byline = `The available datasets and narratives in this group are listed below.`;
-      let website = null;
-      let showDatasets = true;
-      let showNarratives = true;
-      let overview;
-      if (objectKeys.includes("group-overview.md")) {
-        const overviewContent = await this.getAndDecompressObject("group-overview.md");
-        [title, byline, website, showDatasets, showNarratives, overview] = this.parseOverviewMarkdown(overviewContent);
-        // Default show datasets & narratives if not specified in customization
-        if (showDatasets == null) showDatasets = true;
-        if (showNarratives == null) showNarratives = true;
-      }
-
-      return {
-        title: title,
-        byline: byline,
-        website: website,
-        showDatasets: showDatasets,
-        showNarratives: showNarratives,
-        avatar: logoSrc,
-        overview: overview
-      };
-
-    } catch (err) {
-      /* Appropriate fallback if no customised data is available */
-      return {
-        title: `"${this.name}" Nextstrain group`,
-        byline: `The available datasets and narratives in this group are listed below.`,
-        website: null,
-        showDatasets: true,
-        showNarratives: true,
-        error: `Error in custom group info: ${err.message}`
-      };
-    }
-  }
-}
-
-class PrivateS3Source extends S3Source {
-  static visibleToUser(user) { // eslint-disable-line no-unused-vars
-    throw new Error("visibleToUser() must be implemented explicitly by subclasses (not inherited from PrivateS3Source)");
-  }
-  async urlFor(path, method = 'GET') {
-    return S3.getSignedUrl(method === "HEAD" ? "headObject" : "getObject", {
-      Bucket: this.bucket,
-      Key: path
-    });
-  }
-}
-
-class PublicGroupSource extends S3Source {
-  get bucket() { return `nextstrain-${this.name}`; }
-  static isGroup() {
-    return true;
-  }
-}
-
-class PrivateGroupSource extends PrivateS3Source {
-  get bucket() { return `nextstrain-${this.name}`; }
-
-  static visibleToUser(user) {
-    return !!user && !!user.groups && user.groups.includes(this._name);
-  }
-  static isGroup() {
-    return true;
-  }
-}
-
-class BlabSource extends PublicGroupSource {
-  static get _name() { return "blab"; }
-}
-
-class BlabPrivateSource extends PrivateGroupSource {
-  static get _name() { return "blab-private"; }
-}
-
-class InrbDrcSource extends PrivateGroupSource {
-  /* Person to contact for enquiries: Alli Black / James Hadfield */
-  static get _name() { return "inrb-drc"; }
-
-  // INRB's bucket is named differently due to early adoption
-  get bucket() { return "nextstrain-inrb"; }
-}
-
-class NzCovid19PrivateSource extends PrivateGroupSource {
-  /* Person to contact for enquiries: James Hadfield */
-  static get _name() { return "nz-covid19-private"; }
-}
-
-class AllWalesPrivateSource extends PrivateGroupSource {
-  /* Person to contact for enquiries: James Hadfield */
-  static get _name() { return "allwales-private"; }
-}
-
-class NextspainSource extends PublicGroupSource {
-  /* Person to contact for enquiries: James Hadfield */
-  static get _name() { return "nextspain"; }
-}
-
-class SeattleFluSource extends PublicGroupSource {
-  static get _name() { return "seattleflu"; }
-}
-
-class SwissSource extends PublicGroupSource {
-  /* Person to contact for enquiries: Richard Neher / Emma Hodcroft */
-  static get _name() { return "swiss"; }
-}
-
-class COGUKSource extends PublicGroupSource {
-  /* Person to contact for enquiries: Trevor / Emma / James */
-  static get _name() { return "cog-uk"; }
-}
-
-class NGSSASource extends PublicGroupSource {
-  /* Person to contact for enquiries: Richard Neher / Emma Hodcroft */
-  static get _name() { return "ngs-sa"; }
-}
-
-class ECDCSource extends PublicGroupSource {
-  /* Person to contact for enquiries: Richard Neher / Emma Hodcroft */
-  static get _name() { return "ecdc"; }
-}
-
-class IllinoisGagnonPublicSource extends PublicGroupSource {
-  /* Person to contact for enquiries: Thomas Sibley */
-  static get _name() { return "illinois-gagnon-public"; }
-}
-
-class IllinoisGagnonPrivateSource extends PrivateGroupSource {
-  /* Person to contact for enquiries: Thomas Sibley */
-  static get _name() { return "illinois-gagnon-private"; }
-}
-
-class GrubaughLabPrivateSource extends PrivateGroupSource {
-  /* Person to contact for enquiries: James */
-  static get _name() { return "grubaughlab"; }
-}
-
-class NeherLabSource extends PublicGroupSource {
-  /* Person to contact for enquiries: Richard */
-  static get _name() { return "neherlab"; }
-}
-
-class SpheresSource extends PublicGroupSource {
-  /* Person to contaect for enquiries: Trevor */
-  static get _name() { return "spheres"; }
-}
-
-class NIPHSource extends PublicGroupSource {
-  static get _name() { return "niph"; }
-}
-
-class EPICOVIGALSource extends PublicGroupSource {
-  static get _name() { return "epicovigal"; }
-}
-
-class WAPHLSource extends PublicGroupSource {
-  static get _name() { return "waphl"; }
-}
-
-class ILRIPrivateSource extends PrivateGroupSource {
-  static get _name() { return "ilri"; }
-}
-
-class NebraskaDHHSSource extends PublicGroupSource {
-  /* Person to contact: Bryan Temogoh */
-  static get _name() { return "nebraska-dhhs"; }
-}
-
-class PIGIEPrivateSource extends PrivateGroupSource {
-  static get _name() { return "pigie"; }
-}
-
-class ViennaRNASource extends PublicGroupSource {
-  static get _name() { return "ViennaRNA"; }
-  get bucket() { return "nextstrain-viennarna"; }
-}
-
-class SC2ZamPubSource extends PublicGroupSource {
-  static get _name() { return "SC2ZamPub"; }
-  get bucket() { return "nextstrain-sc2zampub"; }
-}
-
-class SC2ZamPrivateSource extends PrivateGroupSource {
-  static get _name() { return "SC2Zam"; }
-  get bucket() { return "nextstrain-sc2zam"; }
-}
-
-class WallauLabPrivateSource extends PrivateGroupSource {
-  static get _name() { return "wallaulab"; }
-}
-
-class NextfluPrivateSource extends PrivateGroupSource {
-  static get _name() { return "nextflu-private"; }
-}
-
-class NcovHKSource extends PublicGroupSource {
-  static get _name() { return "ncovHK"; }
-  get bucket() { return "nextstrain-ncovhk"; }
-}
-
-class DatabiomicsPrivateSource extends PrivateGroupSource {
-  static get _name() { return "databiomics"; }
-}
-
-
-const sources = [
-  CoreSource,
-  CoreStagingSource,
-  CommunitySource,
-  UrlDefinedSource,
-  /* Public nextstrain groups: */
-  BlabSource,
-  SeattleFluSource,
-  NextspainSource,
-  SwissSource,
-  COGUKSource,
-  NGSSASource,
-  ECDCSource,
-  IllinoisGagnonPublicSource,
-  NeherLabSource,
-  SpheresSource,
-  NIPHSource,
-  EPICOVIGALSource,
-  WAPHLSource,
-  ViennaRNASource,
-  SC2ZamPubSource,
-  NebraskaDHHSSource,
-  NcovHKSource,
-  /* Private nextstrain groups: */
-  BlabPrivateSource,
-  NzCovid19PrivateSource,
-  AllWalesPrivateSource,
-  IllinoisGagnonPrivateSource,
-  GrubaughLabPrivateSource,
-  InrbDrcSource,
-  ILRIPrivateSource,
-  PIGIEPrivateSource,
-  SC2ZamPrivateSource,
-  WallauLabPrivateSource,
-  NextfluPrivateSource,
-  DatabiomicsPrivateSource,
-];
-
-const sourceMap = new Map(sources.map(s => [s._name, s]));
-utils.verbose("Sources are:", sourceMap);
-
-module.exports = sourceMap;
